+++ Presentation/Services/RoleService.cs
@@ -1,53 +1,89 @@
-﻿
-using Business.Dtos;
+﻿using Business.Dtos;
 using Business.Factories;
 using Business.Interfaces;
 using Business.Models;
 using Data.Entities;
 using Data.Interfaces;
-using System.Data;
-using System.Linq.Expressions;
 
 namespace Business.Services;
+
 
 public class RoleService(IRoleRepository roleRepository) : IRoleService
 {
     private readonly IRoleRepository _roleRepository = roleRepository;
 
-    public async Task<Role> CreateRoleAsync(RoleRegistrationForm form)
+    public async Task<Result> CreateRoleAsync(RoleRegistrationForm form)
     {
-        var entity = await _roleRepository.GetAsync(x => x.RoleName == form.RoleName);
-        entity ??= await _roleRepository.CreateAsync(RoleFactory.Create(form));
+        if (form == null)
+            return Result.BadRequest("Role registration form is required");
 
-        return RoleFactory.Create(entity);
+        try
+        {
+            if (await _roleRepository.GetAsync(x => ((RoleEntity)x).RoleName == form.RoleName) != null)
+                return Result.AlreadyExists("Role with this email already exists");
+
+            var roleEntity = RoleFactory.Create(form);
+
+            var result = await _roleRepository.CreateAsync(roleEntity);
+            return result != null ? Result.Ok() : Result.BadRequest("Role registration failed");
+        }
+
+        catch (Exception ex)
+        {
+            Console.WriteLine(ex.Message);
+            return Result.BadRequest("Role registration failed");
+
+        }
     }
 
-    public async Task<IEnumerable<Role>> GetAllRoleAsync()
+    public async Task<Result<IEnumerable<Role>>> GetAllRoleAsync()
     {
-        var entities = await _roleRepository.GetAllAsync();
-        var role = entities.Select(RoleFactory.Create);
-        return role ?? Enumerable.Empty<Role>();
+        var roleEntities = await _roleRepository.GetAllAsync();
+        var role = roleEntities?.Select(RoleFactory.Create);
+        return Result<IEnumerable<Role>>.Ok(role);
     }
 
-    public async Task<Role> GetRoleAsync(Expression<Func<RoleEntity, bool>> expression)
+    public async Task<IResult> GetRoleByIdAsync(int id)
     {
-        var entity = await _roleRepository.GetAsync(expression);
-        var role = RoleFactory.Create(entity);
-        return role ?? null!;
+        var roleEntity = await _roleRepository.GetAsync(x => x.Id == id);
+
+        if (roleEntity == null)
+            return Result.NotFound("Role not found");
+
+        var role = RoleFactory.Create(roleEntity);
+        return Result<Role>.Ok(role);
     }
 
-    public async Task<Role> UpdateRoleAsync(RoleUpdateForm form)
+    public async Task<IResult> GetRoleByEmailAsync(string roleName)
     {
-        var entity = await _roleRepository.UpdateAsync(x => x.Id == form.Id, RoleFactory.Create(form));
-        var role = RoleFactory.Create(entity);
-        return role ?? null!;
+        var roleEntity = await _roleRepository.GetAsync(x => x.RoleName == roleName);
+
+        if (roleEntity == null)
+            return Result.NotFound("Role not found");
+
+        var role = RoleFactory.Create(roleEntity);
+        return Result<Role>.Ok(role);
+    }
+    public async Task<IResult> UpdateRoleAsync(int id, RoleUpdateForm updateForm)
+    {
+        var roleEntity = await _roleRepository.GetAsync(x => x.Id == id);
+        if (roleEntity == null)
+            return Result.NotFound("Role not found");
+
+        roleEntity = RoleFactory.Create(roleEntity, updateForm);
+        var result = await _roleRepository.UpdateAsync(x => x.Id == id, roleEntity);
+        return result != null ? Result.Ok() : Result.BadRequest("Role update failed");
     }
 
-    public async Task<bool> DeleteRoleAsync(int id)
+    public async Task<IResult> DeleteRoleAsync(int id)
     {
+        var roleEntity = await _roleRepository.GetAsync(x => x.Id == id);
+        if (roleEntity == null)
+            return Result.NotFound("Role not found");
+
         var result = await _roleRepository.DeleteAsync(x => x.Id == id);
-        return result;
+        return result ? Result.Ok() : Result.BadRequest("Role update failed");
     }
-}
 
+}
